+++ test/Moon.Reflection.Tests/AnonymousObjectTests.cs
@@ -3,13 +3,13 @@ using Xunit;
 
 namespace Moon.Reflection.Tests
 {
-    public class AnonymousTests
+    public class AnonymousObjectTests
     {
         [Fact]
-        public void ConvertingAnonymousToDictionary()
+        public void ConvertingAnonymousObjectToDictionary()
         {
             var anonymous = new { test = "test", test2 = "test2" };
-            var result = Anonymous.ToDictionary(anonymous);
+            var result = AnonymousObject.ToDictionary(anonymous);
 
             result["test"].Should().Be("test");
             result["test2"].Should().Be("test2");